Refac: styled 컴포넌트 분리 및 Edit,Create 공통 컴포넌트 사용
