+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export * from './common';
 export * from './layout';
 export * from './main';
+export * from './quizlet';
