+++ src/components/quizlet/QuestionCardInputField.tsx
@@ -4,26 +4,40 @@ import {
 	UseFieldArrayRemove,
 	UseFormRegister,
 } from 'react-hook-form';
+import styled from '@emotion/styled';
 import { IconButton, TextField, Typography } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
-import { BaseQuizlet, CreateQuizletRequest } from '../../types';
-import StyledBox from './StyledBox';
-import styled from '@emotion/styled';
+import {
+	BaseQuizlet,
+	CreateQuizletRequest,
+	EditQuizletRequest,
+} from '../../types';
+import StyledBox from './styles/StyledBox';
 import { desktopMediaQuery, mobileMediaQuery } from '../../utils/mediaQueries';
 
 export interface QuestionCardInputFieldProps<T extends BaseQuizlet> {
 	index: number;
+	questionNumber: number;
+	listName: 'questionCardList' | 'questionCardListToAdd';
 	register: UseFormRegister<T>;
 	errors: FieldErrors<T>;
 	remove: UseFieldArrayRemove;
 }
 
+function QuestionCardInputField(
+	props: QuestionCardInputFieldProps<CreateQuizletRequest>,
+): JSX.Element;
+function QuestionCardInputField(
+	props: QuestionCardInputFieldProps<EditQuizletRequest>,
+): JSX.Element;
 function QuestionCardInputField({
 	index,
+	questionNumber,
+	listName,
 	register,
 	errors,
 	remove,
-}: QuestionCardInputFieldProps<CreateQuizletRequest>) {
+}: QuestionCardInputFieldProps<BaseQuizlet>) {
 	return (
 		<StyledBox>
 			<ButtonWrapper>
@@ -31,29 +45,29 @@ function QuestionCardInputField({
 					<StyledCloseIcon />
 				</RemoveButton>
 			</ButtonWrapper>
-			<Title>{`Question ${index + 1}`}</Title>
+			<Title>{`Question ${questionNumber}`}</Title>
 			<Wrapper>
 				<TextInput
 					label="문제"
 					variant="outlined"
-					error={!!errors.questionCardList?.[index]?.question}
-					helperText={errors?.questionCardList?.[index]?.question?.message}
-					{...register(`questionCardList.${index}.question`)}
+					error={!!errors[`${listName}`]?.[index]?.question}
+					helperText={errors[`${listName}`]?.[index]?.question?.message}
+					{...register(`${listName}.${index}.question`)}
 				/>
 				<TextInput
 					label="참고 링크"
 					variant="outlined"
-					error={!!errors.questionCardList?.[index]?.link}
-					helperText={errors?.questionCardList?.[index]?.link?.message}
-					{...register(`questionCardList.${index}.link`)}
+					error={!!errors[`${listName}`]?.[index]?.link}
+					helperText={errors[`${listName}`]?.[index]?.link?.message}
+					{...register(`${listName}.${index}.link`)}
 				/>
 				<TextInput
 					label="정답"
 					variant="outlined"
 					multiline
-					error={!!errors.questionCardList?.[index]?.answer}
-					helperText={errors?.questionCardList?.[index]?.answer?.message}
-					{...register(`questionCardList.${index}.answer`)}
+					error={!!errors[`${listName}`]?.[index]?.answer}
+					helperText={errors[`${listName}`]?.[index]?.answer?.message}
+					{...register(`${listName}.${index}.answer`)}
 				/>
 			</Wrapper>
 		</StyledBox>
