+++ src/components/quizlet/index.ts
@@ -1,5 +1,5 @@
 export { default as EditQuizletForm } from './EditQuizletForm';
+export { default as CreateQuizletForm } from './CreateQuizletForm';
 export { default as RemoveQuestionCard } from './RemoveQuestionCard';
 export { default as QuestionCardInputField } from './QuestionCardInputField';
-export { default as EditQuestionCardInputField } from './EditQuestionCardInputField';
 export { default as QuizletInfoInputField } from './QuizletInfoInputField';
