+++ src/components/quizlet/styles/ButtonGroup.tsx
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const ButtonGroup = styled.div`
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	width: 100%;
+	margin-top: 50px;
+	gap: 10px;
+`;
+
+export default ButtonGroup;
