+++ src/components/quizlet/styles/ErrorMessage.tsx
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+const ErrorMessage = styled.div`
+	color: var(--warn-color);
+	font-size: 1.1rem;
+`;
+
+export default ErrorMessage;
