+++ src/components/quizlet/styles/StyledAddIcon.tsx
@@ -1,0 +1,19 @@
+import styled from '@emotion/styled';
+import {
+	desktopMediaQuery,
+	mobileMediaQuery,
+} from '../../../utils/mediaQueries';
+
+import AddIcon from '@mui/icons-material/Add';
+
+const StyledAddIcon = styled(AddIcon)`
+	padding: 0px;
+	${mobileMediaQuery} {
+		font-size: 1.3rem;
+	}
+	${desktopMediaQuery} {
+		font-size: 1.7rem;
+	}
+`;
+
+export default StyledAddIcon;
