+++ src/components/quizlet/styles/StyledBox.tsx
@@ -1,6 +1,9 @@
 import styled from '@emotion/styled';
 import { Box } from '@mui/material';
-import { desktopMediaQuery, mobileMediaQuery } from '../../utils/mediaQueries';
+import {
+	desktopMediaQuery,
+	mobileMediaQuery,
+} from '../../../utils/mediaQueries';
 
 const StyledBox = styled(Box)`
 	display: flex;
