+++ src/components/quizlet/styles/StyledButton.tsx
@@ -1,0 +1,18 @@
+import styled from '@emotion/styled';
+import { Button } from '@mui/material';
+import {
+	desktopMediaQuery,
+	mobileMediaQuery,
+} from '../../../utils/mediaQueries';
+
+const StyledButton = styled(Button)`
+	width: 50%;
+	${mobileMediaQuery} {
+		font-size: 1rem;
+	}
+	${desktopMediaQuery} {
+		font-size: 1.4rem;
+	}
+`;
+
+export default StyledButton;
