+++ src/components/quizlet/styles/StyledForm.tsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+import { Form } from 'react-router-dom';
+
+const StyledForm = styled(Form)`
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+`;
+
+export default StyledForm;
