+++ src/components/quizlet/styles/StyledTitle.tsx
@@ -1,0 +1,20 @@
+import styled from '@emotion/styled';
+import { Typography } from '@mui/material';
+import {
+	desktopMediaQuery,
+	mobileMediaQuery,
+} from '../../../utils/mediaQueries';
+
+const StyledTitle = styled(Typography)`
+	font-weight: 600;
+	word-break: keep-all;
+	color: var(--font-color);
+	${mobileMediaQuery} {
+		font-size: 1.4rem;
+	}
+	${desktopMediaQuery} {
+		font-size: 1.8rem;
+	}
+`;
+
+export default StyledTitle;
